requirements.txt
# Core dependencies for gRPC + REST Gateway
grpcio>=1.60.0
grpcio-tools>=1.60.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
redis>=5.0.0

# Additional dependencies for examples and utilities
requests>=2.31.0